+++ CHANGELOG.md
@@ -6,6 +6,22 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fix: Fix warning in `SelectWithState` component.
 
+## [2.60.2] - 2020-03-26
+
+Fix:
+- More performant `DoubleEntryTable` Component.
+
+## [2.60.1] - 2020-03-25
+
+Fix:
+- Update `CopyIcon`.
+
+## [2.60.0] - 2020-03-19
+
+Features:
+- Add `hidePagination` props to `Carousel`.
+- Create `CameraIcon` component.
+
 ## [2.59.1] - 2020-03-17
 
 Fixes: